Added a comment about an edge case
The method that checks if a room is available purposefully only checks on the floor where lecturers are supposed to be and for rooms of the Lecturer type. If you change that in the room table, this method will produce errors when trying to edit lecturers that stay in that room.